requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
dj-database-url==1.2.0
django-filter==23.3
whitenoise==6.4.0

# AI Integration
google-generativeai==0.3.2

# Email & Communications
sendgrid==6.11.0
celery==5.3.4
redis==5.0.1

# Email Provider SDKs
mailchimp-transactional==1.0.56
sib-api-v3-sdk==7.6.0

# Webhooks & Integrations
requests==2.31.0
cryptography==41.0.7